refactor(editor): Remove i18n and toast usage from setting store
